// superset-frontend/plugins/plugin-chart-table/src/types.ts
export type SortByItem = {
  id: string;
  key: string;
  desc?: boolean;
};

export type SearchOption = {
  value: string;
  label: string;
};